requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
flask-migrate
flask-wtf
flask-bcrypt
python-dotenv
pymysql
cryptography
waitress
